requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
moviepy>=1.0.3
python-dotenv>=1.0.0
pydantic>=2.0.0
google-genai>=1.0.0

